% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.num}
\alias{as.num}
\title{Converts vector of strings to numeric vector}
\usage{
as.num(x, na.strings = "NA")
}
\arguments{
\item{x}{Input vector of strings.}

\item{na.strings}{A string which represents \code{NA}.
Default: "NA"}
}
\value{
A numeric vector
}
\description{
Converts vector of strings to numeric vector
}
\examples{
library(haploR)
as.num(c("1", "2", "X"), na.strings="X")
}
