% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploregStudy.R
\name{getHaploRegStudyList}
\alias{getHaploRegStudyList}
\title{This function queries HaploReg web-based tool 
in order to see a list of GWAS.}
\usage{

  getHaploRegStudyList(url = "http://archive.broadinstitute.org/mammals/haploreg/haploreg.php")
}
\arguments{
\item{url}{A url to HaploReg.
Default: <http://archive.broadinstitute.org/mammals/haploreg/haploreg.php>}
}
\value{
A list of studies. Each study is itself a list
of two: \code{name}, \code{id}.
}
\description{
This function queries HaploReg web-based tool 
in order to see a list of GWAS.
}
\examples{
studies <- getHaploRegStudyList()
studies
}
