% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploregStudy.R
\name{getStudyList}
\alias{getStudyList}
\title{This function queries HaploReg web-based tool 
in order to see a list of GWAS.}
\usage{
getStudyList(url = Haploreg.settings[["study.url"]])
}
\arguments{
\item{url}{A url to HaploReg.
Default: <http://pubs.broadinstitute.org/mammals/haploreg/haploreg.php>
Previously was: <http://archive.broadinstitute.org/mammals/haploreg/haploreg.php>}
}
\value{
A list of studies. Each study is itself a list
of two: \code{name}, \code{id}.
}
\description{
This function queries HaploReg web-based tool 
in order to see a list of GWAS.
}
\examples{
studies <- getStudyList()
studies
}
