% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDlink.R
\name{LDlink.LDmatrix}
\alias{LDlink.LDmatrix}
\title{This function queries HaploReg web-based tool and returns results.}
\usage{
LDlink.LDmatrix(snps, r2d = "r2", population = "ALL", snp.gene.map = NA)
}
\arguments{
\item{snps}{A list of snps (a vector of rsIDs, or a file, one SNP per line).}

\item{r2d}{Show r2 or D'. Default: \code{r2}}

\item{population}{A particular study.
Default: \code{ALL}.}

\item{snp.gene.map}{A list (map) in which a gene is a key.
Example: list("GENE1"=c("SNP1", "SNP2", "SNP3"), "GENE2"=c("SNP4", "SNP5"))
Default: \code{NA}}
}
\value{
A list of two: (1) raw LD matrix and (2) colored (fancy) matrix with LD gradient.
}
\description{
This function queries HaploReg web-based tool and returns results.
}
\examples{
library(haploR)
data <- LDlink.LDmatrix(c("rs10048158","rs4791078"))
head(data)
}
