% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDlink.R
\name{makeStylishLDmatrix}
\alias{makeStylishLDmatrix}
\title{This function queries HaploReg web-based tool and returns results.}
\usage{
makeStylishLDmatrix(ldmat, haploreg.url = Haploreg.settings[["base.url"]])
}
\arguments{
\item{ldmat}{A LDmatrix in which the first column should contain rdIDs}

\item{haploreg.url}{URL to HaploReg project. 
Default: <https://pubs.broadinstitute.org/mammals/haploreg/haploreg.php>}
}
\value{
colored (fancy) matrix with LD gradient (an object of classses \code{datatables}, \code{htmlwidget})
}
\description{
This function queries HaploReg web-based tool and returns results.
}
\examples{
library(haploR)
data <- LDlink.LDmatrix(c("rs10048158","rs4791078"))
head(data)
makeStylishLDmatrix(data$matrix.r2)
}
