% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{queryHaploreg}
\alias{queryHaploreg}
\title{This function queries Haploreg web-based tool and returns results.}
\usage{
queryHaploreg(query, ldThresh = 0.8, ldPop = "EUR",
  epi = "ChromHMM (Core 15-state model)", cons = "siphy", output = "text",
  url = "http://archive.broadinstitute.org/mammals/haploreg/haploreg.php",
  verbose = FALSE)
}
\arguments{
\item{query}{Variant (SNP) code, only one is supported now.}

\item{ldThresh}{LD threshold, r2 (select NA to only show query variants). 
Default: 0.8.}

\item{ldPop}{1000G Phase 1 population for LD calculation. 
Can be: "AFR", "AMR", "ASN". Default: "EUR".}

\item{epi}{Source for epigenomes. 
Default: "ChromHMM (Core 15-state model)".}

\item{cons}{Mammalian conservation algorithm. 
Default: "SiPhy-omega".}

\item{output}{Output mode. Text by default.}

\item{url}{Haplotype url address. 
Default: "http://archive.broadinstitute.org/mammals/haploreg/haploreg.php"}

\item{verbose}{Verbosing output. Default: FALSE.}
}
\value{
A data frame with results.
}
\description{
This function queries Haploreg web-based tool and returns results.
}
\examples{
data <- queryHaploreg("rs10048158")
head(data)
}

