% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulome.R
\name{queryRegulome}
\alias{queryRegulome}
<<<<<<< HEAD
\title{This function queries RegulomeDB web-based tool 
and returns results in a data frame.}
\usage{
queryRegulome(
  query = NULL,
  genomeAssembly = "GRCh37",
  limit = 1000,
  timeout = 100
)
}
\arguments{
\item{query}{Query (a vector of rsIDs or exact one query region in rsid or like "chr1:39492461-39492462").}

\item{genomeAssembly}{Genome assembly built: can be GRCh37 (default) or GRCh38.}

\item{limit}{It controls how many variants will be queried and returned for a large region. It can be a number (1000 by default) or "all". 
Please note that large number or "all" may get yourself hurt because you could get timeout or may even crash the server.}

\item{timeout}{A \code{timeout} parameter for \code{httr::GET}.
Default: 100}
}
\value{
a data frame (table)
OR
a list with the following items:
- guery_coordinates
- features
- regulome_score
- variants
- nearby_snps,
- assembly
}
\description{
This function queries RegulomeDB web-based tool 
=======
\title{This function queries RegulomeDB \url{www.regulomedb.org} web-based tool 
and returns results in a data frame.}
\usage{
queryRegulome(query = NULL, format = "full",
  url = "http://www.regulomedb.org/results", timeout = 100,
  check_bad_snps = TRUE, verbose = FALSE)
}
\arguments{
\item{query}{Query (a vector of rsIDs).}

\item{format}{An output format. 
Can be on of the following: \code{full} - plain text,
\code{bed} -  BED (Browser Extensible Data) format, 
see e.g. <https://genome.ucsc.edu/FAQ/FAQformat.html#format5.1>,
\code{gff} - GFF (General Feature Format), 
see e.g. <https://genome.ucsc.edu/FAQ/FAQformat.html#format3>
Only \code{full} is currently supported.}

\item{url}{Regulome url address. 
Default: <http://www.regulomedb.org/results>}

\item{timeout}{A \code{timeout} parameter for \code{httr::POST}.
Default: 100}

\item{check_bad_snps}{Checks if all SNPs are annotated.
Default: \code{TRUE}}

\item{verbose}{Verbosing output. Default: FALSE.}
}
\value{
A list of two:
(1) a data frame (table) and
(2) a list of bad SNP IDs. 
Bad SNP ID are those IDs 
that were not found in 1000 Genomes 
Phase 1 data
}
\description{
This function queries RegulomeDB \url{www.regulomedb.org} web-based tool 
>>>>>>> 222e12c342c60fd04825117621e09891331b7c32
and returns results in a data frame.
}
\examples{
data <- queryRegulome(c("rs4791078","rs10048158"))
<<<<<<< HEAD
head(data)
=======
head(data[["res.table"]])
head(data[["bad.snp.id"]])
>>>>>>> 222e12c342c60fd04825117621e09891331b7c32
}
