% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDlink.R
\name{LDlink.LDmatrix}
\alias{LDlink.LDmatrix}
\title{This function queries HaploReg web-based tool and returns results.}
\usage{
LDlink.LDmatrix(snps, r2d = "r2", population = "ALL")
}
\arguments{
\item{snps}{A list of snps (a vector of rsIDs, or a file, one SNP per line).}

\item{r2d}{Show r2 or D'. Default: \code{r2}}

\item{population}{A particular genetic population.
Default: \code{ALL}.}
}
\value{
A list of three: (1) raw LD r2 matrix, 
(2) colored (fancy) matrix with LD r2 gradient (an object of classses \code{datatables}, \code{htmlwidget}),
(3) raw LD D-prime matrix.
}
\description{
This function queries HaploReg web-based tool and returns results.
}
\examples{
library(haploR)
data <- LDlink.LDmatrix(c("rs10048158","rs4791078"))
head(data)
}
