% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDlink.R
\name{makeStylishLDmatrix}
\alias{makeStylishLDmatrix}
\title{This function queries HaploReg web-based tool and returns results.}
\usage{
makeStylishLDmatrix(ldmat)
}
\arguments{
\item{ldmat}{A LDmatrix in which the first column should contain rdIDs}
}
\value{
colored (fancy) matrix with LD gradient (an object of classses \code{datatables}, \code{htmlwidget})
}
\description{
This function queries HaploReg web-based tool and returns results.
}
\examples{
library(haploR)
data <- LDlink.LDmatrix(c("rs10048158","rs4791078"))
head(data)
makeStylishLDmatrix(data$matrix.r2)
}
