% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulome.R
\name{queryRegulome}
\alias{queryRegulome}
\title{This function queries Regulome \url{www.regulomedb.org} web-based tool 
and returns results in a data frame.}
\usage{
queryRegulome(query = NULL, format = "full",
  url = "http://www.regulomedb.org/results", verbose = FALSE)
}
\arguments{
\item{query}{Query (a vector of rsIDs).}

\item{format}{An output format. 
Only \code{full} is currently supported.}

\item{url}{Regulome url address. 
Default: <http://www.regulomedb.org/results>}

\item{verbose}{Verbosing output. Default: FALSE.}
}
\value{
A data frame (table) or an object with results similar to 
Regulome uses.
}
\description{
This function queries Regulome \url{www.regulomedb.org} web-based tool 
and returns results in a data frame.
}
\examples{
data <- queryRegulome(c("rs4791078","rs10048158"))
head(data)
}

